Update patch set 14
Patch Set 14:

> Patch Set 12:
> 
> > Patch Set 12:
> > 
> > (2 comments)
> > 
> > > Patch Set 11:
> > > 
> > > Hi Edwin, see below the proposed website and agenda for the Gerrit User Summit 2019 in Gothenburg: can you check the talks and put your name on the one you would be willing to present?
> > 
> > The proposed agenda looks good to me. You can count me in for "Q&A with the Gerrit Code Review maintainers" and "Gerrit Community Roles, how to become a contributor". Alice will do "First-class Gerrit CI integration with Checks". "Prolog-less submit rules, easy and for everyone" would be Patrick's talk, but since he's not at the user summit I could do that talk too.
> 
> Great, will create a detailed page for you to completed, in a follow-up change.
> 
> > I was also thinking about an more interactive session (see "Community Manager - Agenda and Meeting Notes"), but maybe there is no space on the agenda for this?
> 
> Could this be part of the Q&A with the maintainers? Maybe changed to "Q&A with the Community"?

We discussed this in community manager meeting and we would really like to do such an interactive session. We would need a 90min timeslot for this. 

I don't think this is the same as "Q&A with the maintainers" and I wouldn't mix this. But I would be fine to skip "Q&A with the maintainers" or one of the other sessions in favor of the interactive session. Maybe some talks could also be shorter. Do you think we can make space for this interactive session?

Patch-set: 14
Reviewer: Gerrit User 1003935 <1003935@173816e5-2b9a-37c3-8a2e-48639d4f1153>