Add JGit talk
The Gerrit community expressed the wish to be closer aligned with the
JGit community, hence the community managers tried to organize a JGit
talk at one of our user summits. I talked with Ivan Frade (Google) about
this and I'm glad that he is willing to do a JGit talk at this year's
user summit in Sunnyvale. The concrete topic of the talk is still being
discussed but we already want to secure a slot for this talk. Ivan's
preferred day for this talk is Saturday, so I'm taking a slot there.

Details about the talk will be filled in once the topic is decided.

Bug: Issue 10980
Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: If87dcb361acc700d119aa157235a6e65ed3d40a4