Squeeze into the schedule the Zuul talk
The Zuul talk at the Gerrit User Summit Europe was really
interesting. With the forthcoming work by James on building
Gerrit on it, it becomes even more relevant for the Gerrit Community.

Replay the European talk and give an update on where we are
with the adoption in the Gerrit project.

Change-Id: I9e84efe752a1a090df628ed0a61f760b146b4747